requirements.txt
aiogram==3.3
peewee==3.17.1
pillow==10.2.0
googletrans
openpyxl==3.1.2
APScheduler==3.10.4
tortoise-orm==0.20.0